+++ ft_memset.c
@@ -1,6 +1,6 @@
 #include "libft.h"
 
-void *ft_memset(void *b, int c, int len)
+void *ft_memset(void *b, int c, size_t len)
 {
     unsigned char *b2;
     unsigned char c2;
